third_party/blink/web_tests/external/wpt/css/css-values/calc-size/animation/calc-size-height-interpolation.tentative-expected.txt

This is a testharness.js-based test.
Found 64 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [75px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (0) should be [100px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [125px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [150px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [175px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (1) should be [200px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <height> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [225px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [75px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (0) should be [100px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [125px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [150px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [175px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (1) should be [200px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <height> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [225px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [75px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0) should be [100px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [125px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [150px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [175px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (1) should be [200px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [225px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (-0.25) should be [75px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0) should be [100px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.25) should be [125px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.5) should be [150px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (0.75) should be [175px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (1) should be [200px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <height> from neutral to [calc-size(auto, size * 2)] at (1.25) should be [225px]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (-0.25) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (0) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (0.25) should be [25px]
  assert_equals: expected "25px " but got "100px "
[FAIL] CSS Transitions: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (0.5) should be [50px]
  assert_equals: expected "50px " but got "100px "
[FAIL] CSS Transitions: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (0.75) should be [75px]
  assert_equals: expected "75px " but got "100px "
[FAIL] CSS Transitions: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (1.25) should be [125px]
  assert_equals: expected "125px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (-0.25) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (0) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (0.25) should be [25px]
  assert_equals: expected "25px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (0.5) should be [50px]
  assert_equals: expected "50px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (0.75) should be [75px]
  assert_equals: expected "75px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [calc-size(min-content, 0px)] to [calc-size(min-content, size)] at (1.25) should be [125px]
  assert_equals: expected "125px " but got "100px "
[FAIL] CSS Transitions: property <height> from [0] to [calc-size(max-content, size)] at (-0.25) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions: property <height> from [0] to [calc-size(max-content, size)] at (0) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions: property <height> from [0] to [calc-size(max-content, size)] at (0.25) should be [25px]
  assert_equals: expected "25px " but got "100px "
[FAIL] CSS Transitions: property <height> from [0] to [calc-size(max-content, size)] at (0.5) should be [50px]
  assert_equals: expected "50px " but got "100px "
[FAIL] CSS Transitions: property <height> from [0] to [calc-size(max-content, size)] at (0.75) should be [75px]
  assert_equals: expected "75px " but got "100px "
[FAIL] CSS Transitions: property <height> from [0] to [calc-size(max-content, size)] at (1.25) should be [125px]
  assert_equals: expected "125px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [0] to [calc-size(max-content, size)] at (-0.25) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [0] to [calc-size(max-content, size)] at (0) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [0] to [calc-size(max-content, size)] at (0.25) should be [25px]
  assert_equals: expected "25px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [0] to [calc-size(max-content, size)] at (0.5) should be [50px]
  assert_equals: expected "50px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [0] to [calc-size(max-content, size)] at (0.75) should be [75px]
  assert_equals: expected "75px " but got "100px "
[FAIL] CSS Transitions with transition: all: property <height> from [0] to [calc-size(max-content, size)] at (1.25) should be [125px]
  assert_equals: expected "125px " but got "100px "
[FAIL] CSS Animations: property <height> from [0] to [calc-size(max-content, size)] at (-0.25) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Animations: property <height> from [0] to [calc-size(max-content, size)] at (0) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] CSS Animations: property <height> from [0] to [calc-size(max-content, size)] at (0.25) should be [25px]
  assert_equals: expected "25px " but got "100px "
[FAIL] CSS Animations: property <height> from [0] to [calc-size(max-content, size)] at (0.5) should be [50px]
  assert_equals: expected "50px " but got "100px "
[FAIL] CSS Animations: property <height> from [0] to [calc-size(max-content, size)] at (0.75) should be [75px]
  assert_equals: expected "75px " but got "100px "
[FAIL] CSS Animations: property <height> from [0] to [calc-size(max-content, size)] at (1.25) should be [125px]
  assert_equals: expected "125px " but got "100px "
[FAIL] Web Animations: property <height> from [0] to [calc-size(max-content, size)] at (-0.25) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] Web Animations: property <height> from [0] to [calc-size(max-content, size)] at (0) should be [0]
  assert_equals: expected "0px " but got "100px "
[FAIL] Web Animations: property <height> from [0] to [calc-size(max-content, size)] at (0.25) should be [25px]
  assert_equals: expected "25px " but got "100px "
[FAIL] Web Animations: property <height> from [0] to [calc-size(max-content, size)] at (0.5) should be [50px]
  assert_equals: expected "50px " but got "100px "
[FAIL] Web Animations: property <height> from [0] to [calc-size(max-content, size)] at (0.75) should be [75px]
  assert_equals: expected "75px " but got "100px "
[FAIL] Web Animations: property <height> from [0] to [calc-size(max-content, size)] at (1.25) should be [125px]
  assert_equals: expected "125px " but got "100px "
Harness: the test ran to completion.
